medicine_dashboard/requirements.txt
Flask==2.3.2
pyserial==3.5
Werkzeug==2.3.6
